backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
python-dotenv
pdfplumber
python-docx
psycopg2-binary
